requirements.txt
torch>=1.13.0
torchvision>=0.14.0
timm>=0.6.12
opencv-python-headless>=4.7.0
Pillow>=9.4.0
numpy>=1.23.0
matplotlib>=3.6.0
gradio>=4.0.0
